[None][refactor] Move disagg transfer helpers from py_executor into disaggregation/executor
